Add files via upload
The 'Fokker_Plank_Equation.pdf' is a project I did not finish in graduate school. As the reader may verify, I got lost in the theoretical derivations(see the appendix). Consequently, I did not have time to actually write the computational algorithm for modelling the master equation (which was the subject of the paper). Being a physics major, I was at the time new to graduate level math, and even still the attempted derivations in that appendix would likely stump me. But the algorithm presented is so simply compared to other methods in computational analysis that its sort of laughable. But simple is powerful. So too is the random walk model, so  think this could be used to model a number of things.... come the day I find the time to return to it:)